+++ metadata.rb
@@ -2,8 +2,10 @@ name             'ec2configservice'
 maintainer       'base2Services'
 maintainer_email 'a.walker@base2services.com'
 license          'Apache License, Version 2.0'
-description      'Installs/Configures ec2configservice'
+description      'Installs/Configures EC2 Config Service'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
 supports 'windows'
+
+depends 'windows'