+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # 0.1.0
 
-Initial release of ec2configservice
+ * Initial release of ec2configservice
+ * Installation of the latest EC2 config service
+ * Wallpaper Settings config